Serve wrapped PAC file pointing at alpaca
Serve a PAC file that is wrapped such that the `FindProxyForURL`
function returns either `DIRECT` or `PROXY localhost:port` pointing
at alpaca. `DIRECT` is returned when the PAC file that alpaca is using
would return `DIRECT`. `PROXY localhost:port` is returned otherwise.

This allows the OS to be set up to use `http://localhost:3128/alpaca.pac`
as the PAC file to use which will make it use alpaca for any URLs that
alpaca would proxy but otherwise go direct if that is all alpaca is
going to do anyway. This helps with the odd application that doesn't
seem to like going through a proxy at all. If the system were using
the same PAC file that alpaca is using, applications would make
non-proxied direct connections for these same URLs.

If alpaca does not have a PAC file (if the host is on a different
network for instance), then the `FindProxyForURL` function returned on
the `/alpaca.pac` path will just return `DIRECT` always, directing all
traffic away from alpaca.

Issues: #14