Move ProxyFinder.checkForUpdates into request path
Put ProxyFinder.checkForUpdates in the HTTP path by adding a
`WrapHandler` method to check for updates before passing the HTTP
request to the next handler. This makes little difference to the
`ProxyHandler` as it triggers a check for updates on every proxied
request, but we will also want to check for updates on requests to the
`ServeMux` when it is serving content later.

Issues: #14