Log 407 responses from proxy (#85)
When a proxy server responses with 407 Proxy Authentication Required, Alpaca
silently retries the request with authentication. This PR adds logging so that
we know when Alpaca has gone down the code path to authenticate to the proxy.
This makes it easier to see what's going on in cases such as issue #44.